Add 收货地址

+++ rotor-api/src/main/java/com/mouse/api/client/OrderClient.java
@@ -489,7 +489,7 @@ public class OrderClient extends GlobalExceptionHandler implements OrderFeign {
             List<OrderGoodsEntity> orderGoodsEntities = orderGoodsService.findByOrderId(orderId).orElseGet(() -> new ArrayList<>());
             for (OrderGoodsEntity orderGoodsEntity : orderGoodsEntities) {
                 Integer productId = orderGoodsEntity.getProductId();
-                Short number = orderGoodsEntity.getNumber();
+                Integer number = orderGoodsEntity.getNumber();
                 if (productService.reduceStock(productId, number) == 0) {
                     throw new BusinessException("商品货品库存增加失败");
                 }
